adds set modes migration
