+++ modules/@apostrophecms/doc/index.js
@@ -37,8 +37,7 @@ module.exports = {
     self.apos.isNew = await self.detectNew();
     await self.createIndexes();
     self.addLegacyMigrations();
-    self.addCacheFieldMigration();
-    self.addSetPreviousDocsAposModeMigration();
+    self.addMigrations();
   },
   restApiRoutes(self) {
     return {
@@ -1764,38 +1763,14 @@ module.exports = {
           }
         }
       },
-      // Add the "cacheInvalidatedAt" field to the documents that do not have it yet,
-      // and set it to equal doc.updatedAt.
-      setCacheField() {
-        return self.apos.migration.eachDoc({ cacheInvalidatedAt: { $exists: 0 } }, 5, async doc => {
-          await self.apos.doc.db.updateOne({ _id: doc._id }, {
-            $set: { cacheInvalidatedAt: doc.updatedAt }
-          });
-        });
-      },
-      addCacheFieldMigration() {
-        self.apos.migration.add('add-cache-invalidated-at-field', self.setCacheField);
-      },
-
-      async addSetPreviousDocsAposModeMigration () {
-        self.apos.migration.add('set-previous-docs-apos-mode', async () => {
-          await self.apos.doc.db.updateMany({
-            _id: { $regex: ':previous$' },
-            aposMode: { $ne: 'previous' }
-          }, {
-            $set: {
-              aposMode: 'previous'
-            }
-          });
-        });
-      },
 
       async bestAposDocId(criteria) {
         const existing = await self.apos.doc.db.findOne(criteria, { projection: { aposDocId: 1 } });
         return existing?.aposDocId || self.apos.util.generateId();
       },
 
-      ...require('./lib/legacy-migrations')(self)
+      ...require('./lib/legacy-migrations')(self),
+      ...require('./lib/migrations')(self)
     };
   }
 };
